schema changes

+++ src/gateway/internal/handler/user_identity.go
@@ -40,13 +40,13 @@ func (h *Handler) userIdentity(c *gin.Context) {
 	c.Set(userCtx, userId)
 }
 
-func getUserId(c *gin.Context) (int, error) {
+func getUserId(c *gin.Context) (int64, error) {
 	id, ok := c.Get(userCtx)
 	if !ok {
 		return 0, errors.New("user id not found")
 	}
 
-	idInt, ok := id.(int)
+	idInt, ok := id.(int64)
 	if !ok {
 		return 0, errors.New("user id is of invalid type")
 	}
